feat: 관리자 기능 추가 - 동아리 승인 시 동아리 생성 - 동아리 정보에 최초로 동아리 장 등록

+++ src/main/java/environment/project/dto/ClubInfoDTO.java
@@ -6,7 +6,7 @@ import java.util.Date;
 
 @Data
 public class ClubInfoDTO {
-    private int idx;
+    private Long idx;
     private int clubUserGrade;
     private boolean clubUsed;
     private Date clubJoinDate;
@@ -14,6 +14,6 @@ public class ClubInfoDTO {
     private Date clubOutDate;
 
     // JoinData
-    private int clubNo;
-    private int userNo;
+    private Long clubNo;
+    private Long userNo;
 }
